+++ src/lib.rs
@@ -22,4 +22,5 @@ pub mod transit_realtime {
 pub mod context;
 pub mod gtfs_rt;
 pub mod server;
+pub mod siri_model;
 pub mod stoppoints_discovery;
