+++ src/stoppoints_discovery.rs
@@ -1,70 +1,8 @@
 use actix_web::{Json, Query, Result, State};
 use crate::context::Context;
 use gtfs_structures;
+use siri_model::{AnnotatedStopPoint, Siri, StopPointsDelivery};
 use std::borrow::Borrow;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ErrorCondition {}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct Line {
-    pub line_ref: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Location {
-    pub longitude: f64,
-    pub latitude: f64,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct AnnotatedStopPoint {
-    pub stop_point_ref: String,
-    pub stop_name: String,
-    pub lines: Vec<Line>,
-    pub location: Location,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct StopPointsDelivery {
-    pub version: String,
-    pub response_time_stamp: String,
-    pub status: bool,
-    pub error_condition: Option<ErrorCondition>,
-    pub annotated_stop_point: Vec<AnnotatedStopPoint>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct Siri {
-    pub stop_points_delivery: StopPointsDelivery,
-}
-
-impl AnnotatedStopPoint {
-    fn from(stop: &gtfs_structures::Stop, context: &Context) -> Self {
-        let lines = context
-            .lines_of_stops
-            .get(&stop.id)
-            .unwrap_or(&std::collections::HashSet::new())
-            .iter()
-            .map(|route_id| Line {
-                line_ref: route_id.to_owned(),
-            }).collect();
-
-        Self {
-            stop_point_ref: stop.id.to_owned(),
-            stop_name: stop.name.to_owned(),
-            lines,
-            location: Location {
-                longitude: stop.longitude,
-                latitude: stop.latitude,
-            },
-        }
-    }
-}
 
 #[derive(Deserialize)]
 pub struct Params {
@@ -96,7 +34,15 @@ fn bounding_box_matches(
         && stop.latitude <= max_lat
 }
 
-pub fn stoppoints_discovery((state, query): (State<Context>, Query<Params>)) -> Result<Json<Siri>> {
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct Response {
+    pub siri: Siri,
+}
+
+pub fn stoppoints_discovery(
+    (state, query): (State<Context>, Query<Params>),
+) -> Result<Json<Response>> {
     let stops = &state.gtfs.stops;
 
     let request = query.into_inner();
@@ -113,13 +59,15 @@ pub fn stoppoints_discovery((state, query): (State<Context>, Query<Params>)) -> 
         .map(|stop| AnnotatedStopPoint::from(stop.borrow(), &state))
         .collect();
 
-    Ok(Json(Siri {
-        stop_points_delivery: StopPointsDelivery {
-            version: "2.0".to_string(),
-            response_time_stamp: chrono::Utc::now().to_rfc3339(),
-            annotated_stop_point: filtered,
-            error_condition: None,
-            status: true,
+    Ok(Json(Response {
+        siri: Siri {
+            stop_points_delivery: StopPointsDelivery {
+                version: "2.0".to_string(),
+                response_time_stamp: chrono::Utc::now().to_rfc3339(),
+                annotated_stop_point: filtered,
+                error_condition: None,
+                status: true,
+            },
         },
     }))
 }
